LayoutTests/imported/w3c/web-platform-tests/editing/other/plain-text-copy-paste-of-paragraph-ending-with-non-layed-out-content-expected.txt

line 1

line 2

line 3

test;


FAIL The extra line break is missing after the paragraph that ends with a comment. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL The extra line break is missing after the paragraph that ends with a display:none span. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL The extra line break is missing after the paragraph that ends with a hidden span. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL The extra line break is missing after the paragraph that ends with a meta tag. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL The extra line break is missing after the paragraph that ends with a style tag. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL The extra line break is missing after the paragraph that ends with a base tag. assert_equals: expected "line 1\n\nline 2\n\nline 3" but got ""
FAIL Line break should not be present before semicolon after block elements styled as inline-flex. assert_equals: expected "test;" but got ""
FAIL Line break should not be present before semicolon after block elements styled as inline-block. assert_equals: expected "test;" but got ""
FAIL Line break should not be present before semicolon after block elements styled as inline-grid. assert_equals: expected "test;" but got ""
